requirements-test.txt
coverage
coveralls
flake8>=2.1.0
tox>=2.1.1
django-autoslug-iplweb
pytest
pytest-django
